+++ tests/test_fin_evaluator.py
